requirements.txt
aiogram==2.11.2
aiohttp>=3.7.3
aioredis==1.3.1
python-dotenv~=0.15.0

SQLAlchemy~=1.3.22
schedule~=0.6.0
APScheduler~=3.6.3
emoji~=0.6.0